Restful APIs and Flask Code/Restful APIs and Flask Code/sentimentRedditAnalyzer/requirements.txt
alembic==1.14.0
altair==5.0.1
argcomplete==2.1.2
attrs==24.3.0
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
commonmark==0.9.1
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
filelock==3.16.1
Flask==2.2.5
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
nltk==3.9.1
nox==2022.11.21
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
protobuf==5.29.3
psycopg2==2.9.10
pyarrow==18.1.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.4
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
render-cli==0.2.1
requests==2.32.3
rich==12.6.0
rpds-py==0.22.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.37
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.3.0
virtualenv==20.28.1
watchdog==6.0.0
websocket-client==1.8.0
Werkzeug==2.2.3
wordcloud==1.9.4
